reworked ProcessOutputReader.cs ( 1/x )

+++ src/WGet.NET/HelperClasses/ProcessOutputReader.cs
@@ -3,7 +3,10 @@
 // https://github.com/basicx-StrgV/                 //
 //--------------------------------------------------//
 using System.Collections.Generic;
+using System.Data.Common;
+using System.Diagnostics;
 using System.Text;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace WGetNET.HelperClasses
 {
@@ -25,23 +28,23 @@ namespace WGetNET.HelperClasses
         /// </returns>
         public static List<WinGetPackage> ToPackageList(string[] output)
         {
-            //Get top line index.
+            //Get top line start.
             //The array should always contain this line.
             //If it dose not contain this line the resulting out of range exception,
             //that will be thrown later, will be catched in the calling method.
-            int labelLine = ArrayManager.GetEntryContains(output, "------") - 1;
+            var labelLine = ArrayManager.GetEntryContains(output, "------") - 1;
 
-            //Get start indexes of each tabel colum
-            //(The line starts wich the name followed by the id.
-            //The index for the name is always 0 an therfor it is not configured here.)
-            int idStartIndex = GetColumnStartIndex(output[labelLine], 2);
-            int versionStartIndex = GetColumnStartIndex(output[labelLine], 3);
-            int extraInfoStartIndex = GetColumnStartIndex(output[labelLine], 4);
+            if(labelLine == -1)
+            {
+                return new List<WinGetPackage>();
+            }
+
+            var columns = GetColumnInformation(output[labelLine]);
 
             //Remove unneeded output Lines
             output = ArrayManager.RemoveRange(output, 0, labelLine + 2);
 
-            return CreatePackageListFromOutput(output, idStartIndex, versionStartIndex, extraInfoStartIndex);
+            return CreatePackageListFromOutput(output, columns);
         }
 
         /// <summary>
@@ -50,42 +53,108 @@ namespace WGetNET.HelperClasses
         /// <param name="output">
         /// The <see langword="array"/> containing the output.
         /// </param>
-        /// <param name="idStartIndex">
-        /// The <see cref="System.Int32"/> representing the start index of the id.
-        /// </param>
-        /// <param name="versionStartIndex">
-        /// The <see cref="System.Int32"/> representing the start index of the version.
-        /// </param>
-        /// <param name="extraInfoStartIndex">
-        /// The <see cref="System.Int32"/> representing the start index of the extra information.
+        /// <param name="columns">
+        /// The <see cref="ColumnInfoList"/> representing an IList{<see cref="ColumnInfo"/>}
         /// </param>
         /// <returns>
         /// A <see cref="System.Collections.Generic.List{T}"/> of <see cref="WGetNET.WinGetPackage"/>'s.
         /// </returns>
-        private static List<WinGetPackage> CreatePackageListFromOutput(
-            string[] output, int idStartIndex, int versionStartIndex, int extraInfoStartIndex)
+        private static List<WinGetPackage> CreatePackageListFromOutput(string[] output, ColumnInfoList columns)
         {
-            List<WinGetPackage> resultList = new List<WinGetPackage>();
+            var resultList = new List<WinGetPackage>();
 
-            for (int i = 0; i < output.Length; i++)
+            for (var line = 0; line < output.Length; line++)
             {
-                // [var1..var2] : selects the index range from var1 to var2
-                // (eg. if var1 is 2 and var2 is 5, the selectet index range will be [2, 3, 4])
-                if (output[i].Length >= extraInfoStartIndex) {
-                    try
+                try
+                {
+                    var package = new WinGetPackage()
                     {
-                        resultList.Add(
-                            new WinGetPackage()
-                            {
-                                PackageName = output[i][0..idStartIndex].Trim(),
-                                PackageId = output[i][idStartIndex..versionStartIndex].Trim(),
-                                PackageVersion = output[i][versionStartIndex..extraInfoStartIndex].Trim()
-                            });
-                    }
-                    catch
+                        PackageName = output[line][columns[0].StartIndex..columns[0].End].Trim(),
+                        PackageId = output[line][columns[1].StartIndex..columns[1].End].Trim(),
+                        PackageVersion = output[line][columns[2].StartIndex..columns[2].End].Trim(),
+                        Source = output[line][columns[^1].StartIndex..].Trim()
+                    };
+
+                    if (!package.IsEmpty)
                     {
-                        //Invalid entrys can be skiped
+                        resultList.Add(package);
                     }
+                }
+                catch
+                {
+                    //Invalid entrys can be skiped
+                }
+            }
+
+            return resultList;
+        }
+
+        /// <summary>
+        /// Converts a <see cref="System.Collections.Generic.List{T}"/> 
+        /// of output lines from a winget process to a list of <see cref="WinGetUpdatePackage"/>'s.
+        /// </summary>
+        /// <param name="output">
+        /// A <see cref="System.Collections.Generic.List{T}"/> of output lines from a winget process.
+        /// </param>
+        /// <returns>
+        /// A <see cref="System.Collections.Generic.List{T}"/> of <see cref="WinGetUpdatePackage"/>'s.
+        /// </returns>
+        public static List<WinGetUpdatePackage> ToUpdatePackageList(string[] output)
+        {
+            //Get top line start.
+            var labelLine = ArrayManager.GetEntryContains(output, "------") - 1;
+
+            if (labelLine < 0)
+            {
+                // this will happen, when there are no updates
+                return new List<WinGetUpdatePackage>();
+            }
+
+            var columns = GetColumnInformation(output[labelLine]);
+
+            //Remove unneeded output Lines
+            output = ArrayManager.RemoveRange(output, 0, labelLine + 2);
+
+            return CreateUpdatePackageListFromOutput(output, columns);
+        }
+
+        /// <summary>
+        /// Creates an updatable package list from output.
+        /// </summary>
+        /// <param name="output">
+        /// The <see langword="array"/> containing the output.
+        /// </param>
+        /// <param name="columns">
+        /// The <see cref="ColumnInfoList"/> representing an IList{<see cref="ColumnInfo"/>}
+        /// </param>
+        /// <returns>
+        /// A <see cref="System.Collections.Generic.List{T}"/> of <see cref="WinGetUpdatePackage"/>'s.
+        /// </returns>
+        private static List<WinGetUpdatePackage> CreateUpdatePackageListFromOutput(string[] output, ColumnInfoList columns)
+        {
+            var resultList = new List<WinGetUpdatePackage>();
+
+            for (var line = 0; line < output.Length; line++)
+            {
+                try
+                {
+                    var update = new WinGetUpdatePackage()
+                    {
+                        PackageName = output[line][columns[0].StartIndex..columns[0].End].Trim(),
+                        PackageId = output[line][columns[1].StartIndex..columns[1].End].Trim(),
+                        PackageVersion = output[line][columns[2].StartIndex..columns[2].End].Trim(),
+                        UpdateVersion = output[line][columns[3].StartIndex..columns[3].End].Trim(),
+                        Source = output[line][columns[4].StartIndex..].Trim()
+                    };
+
+                    if (!update.IsEmpty)
+                    {
+                        resultList.Add(update);
+                    }
+                }
+                catch
+                {
+                    //Invalid entrys can be skiped
                 }
             }
 
@@ -104,7 +173,7 @@ namespace WGetNET.HelperClasses
         /// </returns>
         public static List<WinGetSource> ToSourceList(string[] output)
         {
-            //Get top line index.
+            //Get top line start.
             //The array should always contain this line.
             //If it dose not contain this line the resulting out of range exception,
             //that will be thrown later, will be catched in the calling method.
@@ -112,7 +181,7 @@ namespace WGetNET.HelperClasses
 
             //Get start indexes of each tabel colum
             //(The line starts wich the name followed by the id.
-            //The index for the name is always 0 an therfor it is not configured here.)
+            //The start for the name is always 0 an therfor it is not configured here.)
             int urlStartIndex = GetColumnStartIndex(output[labelLine], 2);
 
             //Remove unneeded output Lines
@@ -155,7 +224,7 @@ namespace WGetNET.HelperClasses
         /// The <see langword="array"/> containing the output.
         /// </param>
         /// <param name="urlStartIndex">
-        /// The <see cref="System.Int32"/> representing the start index of the url.
+        /// The <see cref="System.Int32"/> representing the start start of the url.
         /// </param>
         /// <returns>
         /// A <see cref="System.Collections.Generic.List{T}"/> of <see cref="WGetNET.WinGetSource"/>'s.
@@ -166,10 +235,10 @@ namespace WGetNET.HelperClasses
 
             for (int i = 0; i < output.Length; i++)
             {
-                // [var1..var2] : selects the index range from var1 to var2.
-                // (eg. if var1 is 2 and var2 is 5, the selectet index range will be [2, 3, 4])
-                // [var1..] : selects the index range from var1 to the end.
-                // (eg. if var1 is 2 and the last index is 5, the selectet index range will be[2, 3, 4, 5])
+                // [var1..var2] : selects the start range from var1 to var2.
+                // (eg. if var1 is 2 and var2 is 5, the selectet start range will be [2, 3, 4])
+                // [var1..] : selects the start range from var1 to the end.
+                // (eg. if var1 is 2 and the last start is 5, the selectet start range will be[2, 3, 4, 5])
                 resultList.Add(
                     new WinGetSource()
                     {
@@ -181,9 +250,35 @@ namespace WGetNET.HelperClasses
 
             return resultList;
         }
-    
+
+        private static ColumnInfoList GetColumnInformation(string line)
+        {
+            var columns = line.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+            var infos = new ColumnInfoList();
+
+            for (var columnIndex = 0; columnIndex < columns.Length; columnIndex++)
+            {
+                var column = columns[columnIndex];
+                var start = line.IndexOf(column);
+
+                if (start >= 0)
+                {
+                    var next = columnIndex + 1 == columns.Length ? line.Length : line.IndexOf(columns[columnIndex + 1]);
+
+                    infos.Add(new ColumnInfo()
+                    {
+                        Name = column,
+                        StartIndex = start,
+                        End = next
+                    });
+                }
+            }
+
+            return infos;
+        }
+
         /// <summary>
-        /// Gets the start index of a column in the given line.
+        /// Gets the start start of a column in the given line.
         /// </summary>
         /// <param name="line">
         /// A <see cref="System.String"/> representing the line that contains the column names.
@@ -192,11 +287,13 @@ namespace WGetNET.HelperClasses
         /// A <see cref="System.Int32"/> representing the column to look for.
         /// </param>
         /// <returns>
-        /// A <see cref="System.Int32"/> representing the start index of the column 
+        /// A <see cref="System.Int32"/> representing the start start of the column 
         /// or -1 if the column was not found.
         /// </returns>
         private static int GetColumnStartIndex(string line, int column)
         {
+            var columns = line.Split(' ', StringSplitOptions.RemoveEmptyEntries);
+
             int currentColumn = 0;
             bool checkForColumn = true;
             for (int i = 0; i < line.Length; i++)
